##
##  
##   $Id$
##
##
\name{RUnit}
\alias{RUnit}
\title{RUnit - Package Description}
\description{This package models the common Unit Test framework for R
and provides functionality to track results of test case execution and
generate a summary report. It also provides tools for code inspection
and thus for test case coverage analysis.
}

\references{useR! 2004 Vienna}
\author{Thomas K\"onig, Klaus J\"unemann \& Matthias Burger}

\seealso{See Also as \code{\link{~~fun~~}}, ~~~
 or\code{\link{}} }

\examples{
##---- Should be DIRECTLY executable !! ----
}
\keword{programming}
\keyword{RUnit}
