%%  RUnit : A unit test framework for the R programming language
%%  Copyright (C) 2003-2009  Thomas Koenig, Matthias Burger, Klaus Juenemann
%%
%%  This program is free software; you can redistribute it and/or modify
%%  it under the terms of the GNU General Public License as published by
%%  the Free Software Foundation; version 2 of the License.
%%
%%  This program is distributed in the hope that it will be useful,
%%  but WITHOUT ANY WARRANTY; without even the implied warranty of
%%  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
%%  GNU General Public License for more details.
%%
%%  You should have received a copy of the GNU General Public License
%%  along with this program; if not, write to the Free Software
%%  Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA

%%  $Id: RUnit-internal.Rd,v 1.5 2009/04/16 09:13:31 burgerm Exp $

\name{RUnit-internal}

\alias{includeTracker}
\alias{plotConnection.trackInfo}
\alias{writeBeginBody}
\alias{writeBeginHead}
\alias{writeBeginHtml}
\alias{writeBeginTag}
\alias{writeBeginTable}
\alias{writeCR}
\alias{writeEndBody}
\alias{writeEndHead}
\alias{writeEndHtml}
\alias{writeEndTable}
\alias{writeEndTag}
\alias{writeHtmlEnd}
\alias{writeHtmlHeader}
\alias{writeHtmlSection}
\alias{writeHtmlSep}
\alias{writeImage}
\alias{writeLi}
\alias{writeLinkRef}
\alias{writeLink}
\alias{writeP}
\alias{writeRaw}
\alias{writeRawCR}
\alias{writeTableRow}
\alias{writeTitle}

\title{Internal functions}
\description{
  Helper functions and private methods not intended for direct use.
}
\keyword{internal}
