##
##  
##   $Id$
##
##
\name{RUnit}
\alias{RUnit}
\title{RUnit - Package Description}
\description{This package models the common Unit Test framework for R
and provides functionality to track results of test case execution and
generate a summary report. It also provides tools for code inspection
and thus for test case coverage analysis.
}
\section{Class layout}{
Objects can be created by calls of the form \code{new("Basis", ...)}.
    ~~ describe objects here ~~ 
}
\section{Methods}{
  \describe{
    \item{\code{SVMState}:}{Object of class \code{""} ~~ }
  }
}

\references{useR! 2004 Vienna}
\author{RUnit devloper team}
\note{This package is under constant development, all classes and
methods are to be considered unfinalized and may change without prior notice.}

\seealso{See Also as \code{\link{~~fun~~}}, ~~~
 or\code{\link{}} }

\examples{
##---- Should be DIRECTLY executable !! ----
}
\keword{programming}
\keyword{error}
\keyword{RUnit}